+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Key: [+] added, [-] removed, [~] modified, [!] important.
 ## Y-1.0.0
 
 + [!] Completely rebranded the bot into something more personable instead of a random anecdotal name that only a few people will understand.
++ [+] Put in place measures to allow for a web console to edit settings and also see status updates.
++ [~] Updated the system clock to keep time in a more meaningful way that can be used and isn't pre-formatted.
++ [~] Updated exception util to print the full stack trace so that debugging can be done more efficiently.
 + [~] Fixed a number of old bugs the haunted the previous versions.
 
 ## 3.4.0